Moved files to pages folder and render all singed up users to connect file
