+++ client/src/components/Explore.js
@@ -1,18 +1,11 @@
 import * as React from 'react';
-import { Link } from 'react-router-dom';
 import Typography from '@mui/joy/Typography';
 import DescriptionLayout from '../containers/DescriptionLayout';
 import ExploreIcon from '@mui/icons-material/Explore';
-import AspectRatio from '@mui/joy/AspectRatio';
+
 import Box from '@mui/joy/Box';
 import Sheet from '@mui/joy/Sheet';
-import Card from '@mui/joy/Card';
-import CardOverflow from '@mui/joy/CardOverflow';
-import CardCover from '@mui/joy/CardCover';
 
-import articleOneImage from '../assets/images/article-one.webp';
-import articleTwoImage from '../assets/images/article-two.jpeg';
-import articleThreeImage from '../assets/images/article-three.jpg';
 import opportunitiesImage from '../assets/images/opportunities.jpg';
 import successImage from '../assets/images/success.jpg';
 
@@ -84,150 +77,6 @@ export const Explore = () => {
       </Typography>
     </DescriptionLayout>
 
-      <Sheet
-          sx={{
-            display: { xs: 'none', sm: 'initial' },
-            borderColor: 'neutral.outlinedBorder',
-          }}
-        >
-          <Box sx={{ p: 2, display: 'flex', alignItems: 'center' }}>
-          <Typography color="primary" fontSize="lg" fontWeight="lg">
-            Explore More Article
-          </Typography>
-          </Box>
-          <Box
-            sx={{
-              display: 'grid',
-              gridTemplateColumns: 'repeat(auto-fit, minmax(240px, 1fr))',
-              gap: 2,
-            }}
-            >
-            {/* Article One */}
-            <Card variant="outlined">
-              <CardOverflow
-                sx={{
-                  borderBottom: '.5px solid',
-                  borderColor: 'neutral.outlinedBorder',
-                }}
-              >
-                <AspectRatio ratio="16/9" color="primary">
-                <CardCover
-                sx={{
-                  backgroundImage: `url(${articleOneImage})`,
-                  backgroundSize: 'cover',
-                  transition: 'transform 0.3s ease',
-                      '&:hover': {
-                     transform: 'scale(1.05)',
-                      },
-                }}
-                >
-                 <Link
-        to="https://www.news-herald.com/2023/04/11/alumni-should-have-active-role-in-inspiring-current-hs-sports-programs-opinion/"
-        target="_blank"
-        rel="noopener noreferrer"
-      >
-        click
-      </Link>
-                </CardCover>
-                </AspectRatio>
-              </CardOverflow>
-              <Box sx={{ display: 'flex', alignItems: 'center' }}>
-                <Box sx={{ flex: 1 }}>
-                  <Typography>Alumni should have active role in inspiring current HS sports programs | Opinion</Typography>
-                  <Typography level="body3" mt={0.5}>
-                    Created By: Chris Lillstrung
-                  </Typography>
-                  <Typography level="body3" mt={0.5}>
-                    Created: Saturday, July 15th 2023
-                  </Typography>
-                </Box>
-              </Box>
-            </Card>
-
-            {/* Article Two */}
-            <Card variant="outlined">
-              <CardOverflow
-                sx={{
-                  borderBottom: '.5px solid',
-                  borderColor: 'neutral.outlinedBorder',
-                }}
-              >
-                <AspectRatio ratio="16/9" color="primary">
-                <CardCover
-                sx={{
-                  backgroundImage: `url(${articleTwoImage})`,
-                  backgroundSize: 'cover',
-                  transition: 'transform 0.3s ease',
-                      '&:hover': {
-                     transform: 'scale(1.05)',
-                      },
-                }}
-                >
-                  <Link
-        to="https://www.jamesgmartin.center/2023/05/undoing-diversity-equity-and-inclusion-requires-alumni-effort/" 
-        target="_blank"
-        rel="noopener noreferrer"
-      >
-        click
-      </Link>
-                 
-                </CardCover>
-                </AspectRatio>
-              </CardOverflow>
-              <Box sx={{ display: 'flex', alignItems: 'center' }}>
-                <Box sx={{ flex: 1 }}>
-                  <Typography>Undoing “Diversity, Equity, and Inclusion” Requires Alumni Effort</Typography>
-                  <Typography level="body3" mt={0.5}>
-                    Created By: Garland Tucker
-                  </Typography>
-                  <Typography level="body3" mt={0.5}>
-                    Created: Monday, May 29th 2023
-                  </Typography>
-                </Box>
-              </Box>
-            </Card>
-
-            {/* Article Three */}
-            <Card variant="outlined">
-              <CardOverflow
-                sx={{
-                  borderBottom: '.5px solid',
-                  borderColor: 'neutral.outlinedBorder',
-                }}
-              >
-                <AspectRatio ratio="16/9" color="primary">
-                <CardCover
-                sx={{
-                  backgroundImage: `url(${articleThreeImage})`,
-                  backgroundSize: 'cover',
-                  transition: 'transform 0.3s ease',
-                      '&:hover': {
-                     transform: 'scale(1.05)',
-                      },
-                }}
-                >
-                  <Link
-                  to="https://www.bu.edu/articles/2023/5-tips-for-life-after-college-a-guide-to-living-life-as-an-alumni/" 
-                  target="_blank" 
-                  rel="noopener noreferrer"
-                  > click</Link>
-                </CardCover>
-                </AspectRatio>
-              </CardOverflow>
-              <Box sx={{ display: 'flex', alignItems: 'center' }}>
-                <Box sx={{ flex: 1 }}>
-                  <Typography>5 Tips for Life After College: A Guide To Living Life As An Alumni</Typography>
-                  <Typography level="body3" mt={0.5}>
-                    Created By: Jada Warmington
-                  </Typography>
-                  <Typography level="body3" mt={0.5}>
-                    Posted: Tuesday, May 16th 2023
-                  </Typography>
-                </Box>
-              </Box>
-            </Card>
-          </Box>
-        </Sheet>
     </div>
   );
 }
