+++ client/src/components/index.js
@@ -1,9 +1,5 @@
 export * from "./Navbar";
 export * from "./Footer";
-export * from "./Login";
-export * from "./Signup";
-export * from "./Dashboard";
 export * from "./Explore";
 export * from "./Connect";
-export * from "./Profile";
-export * from "./UsersProfile"
+export * from './Header';
