+++ client/src/pages/Login.js
@@ -21,8 +21,8 @@ import { useMutation } from '@apollo/client';
 import { LOGIN } from '../utils/mutations';
 import Auth from '../utils/auth';
 import { Grid } from '@mui/material';
-import UserDashboard from '../pages/UserDashboard';
 import { Link } from "react-router-dom";
+import Dashboard from "./Dashboard";
 
 const ColorSchemeToggle = ({ onClick, ...props }) => {
   const { mode, setMode } = useColorScheme();
@@ -125,7 +125,7 @@ const ColorSchemeToggle = ({ onClick, ...props }) => {
 
     // if user is logged in the render dashboard 
     if (isLoggedIn) {
-      return <UserDashboard />;
+      return <Dashboard />;
     }
