--- client/src/pages/UserDashboard.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import Dashboard from '../components/Dashboard'
-
-const UserDashboard = () => {
-  return (
-    <div>
-        <Dashboard/>
-    </div>
-  )
-}
-
-export default UserDashboard
