+++ client/src/utils/queries.js
@@ -1,5 +1,16 @@
 import { gql } from '@apollo/client';
 
+export const QUERY_PROFILES = gql`
+  query allProfiles {
+    users {
+      _id
+      firstname
+      lastname
+      email
+    }
+  }
+`;
+
 export const GET_USERS = gql`
   query getUsers {
     users {
